Add infrastructure to start using fakes
In this commit the infrastructure to add fakes is added.
Now we can create fake implementations of our objects in
`src/test/unit/test_doubles/path/to/file/realfilename_fake.c`

When we use them we need to add the following to the makefile:

```
shmem_fake.t: $(TEST_DOUBLE_DIR)/backend/storage/ipc/shmem_fake.o
```

This will include the fake instead of the real implementation of
the specific object.

As a proof of concept a shmem_fake was added with some tests that cover
both the fake and the real implementation.

---

Fakes are a different flavor of test doubles. Fakes and Mocks are Test
Doubles are used for different purposes.
The Fake object is usually used when we want the test double to mimics
the behavior of object that we are replacing. This new object will
respect the interface and will behave like the real object, but usually
is implemented in a much more simple way and should not be used in a
production environment.
An example of a common use of a Fake is a InMemoryDatabase.

In order to ensure that the Fakes are kept up to date we need to ensure
that the tests we write for the real implementation are also running
against the Fake, so that we are sure our Fake is still an acurate
representation of the real object.

The Mock object in the other hand functionally they are not smart. Their
objective is to ensure that when you call the test double the interface
is the same as the real object, but we control what is being returned at
each call, and in the end they check that the object was called with the
parameters that was expected.

As per the explanation above the cost of writting a Fake is higher then
writting a Mock in the short term. But in the long run if we need to use
the same Mock object over and over again we will realize that eventually
a Fake would have spared us some Test Code. In our example the Shared
Memory object is a object that will be used in multiple instances of our
tests and it make it easier to create a Fake that is able to replace the
Real object and maintain the behavior.