WIP: Patch to port lwlock tests into 9.4
Correction on the patch that was created for 9.3 that will make
these tests work with 9.4 changes on lwlock

Removed internal structures that were created on the tests
to mimic the ones that were in lwlock.c file, because
in 9.4 the structures were surfaced to the memutils.h now being
public
Removed the tests for the function `NumLWLocks` as the function
is no longer part of the public API
Changed the tests that call `LWLockAcquire` and `LWLockRelease`
because the public API changed

Simplified the helper functions that were to deep in the code
implementation.

Due to a change in `CreateLWLocks` now some more memory is allocated
in a specific context, so we had to initialize the memory contexts in
before the tests. This could be prevented if we had mocked/faked out
the `MemoryContextAlloc` function.

None of the logic of the tests had to change since the behavior as
of the methods from an outside prespective did not change.
Nevertheless some adaptation were always expected due to change in
the public API.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>